refactor: adjust path to component folder
