+++ blockchain/urls.py
@@ -1,7 +1,7 @@
 from django.urls import path 
 
-from blockchain.views.registry import RegistryView
-from blockchain.views.transactions import  TransactionsView
+from blockchain.components.registry.registry_view import RegistryView
+from blockchain.components.transactions.transactions_view import  TransactionsView
 from blockchain.views.mine import MineView
 from blockchain.views.nodes import NewNodeView, RegisterNodeView, BlockSyncView
 
@@ -9,7 +9,7 @@ urlpatterns = [
     path("registry/", RegistryView.as_view()),
     path("registry/<int:id_>", RegistryView.as_view()),
 
-    path("pending/", TransactionsView.as_view()),
+    path("transaction/", TransactionsView.as_view()),
     
     path("mine/", MineView.as_view()),
