+++ blockchain/views/mine.py
@@ -3,7 +3,7 @@ import json, logging
 from rest_framework.views import APIView
 
 
-from ..services.registry_service import RegistryService
+from ..components.registry.registry_service import RegistryService
 
 from blockchain.libraries.peers_manager import PeersManager
 from ..libraries.blockchain import Blockchain
